part4 - 4.17 - 4.23

+++ part4/blog/controllers/blogs.js
@@ -1,39 +1,39 @@
-const express = require('express');
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../middleware/handleToken')
 
 blogRouter.get('/', async (request, response, next) => {
   try {
-    const blog = await Blog.find({}).populate('user')
-    if (blog) {
-      response.json(blog)
-    } else {
-      response.status(404).end()
-    }
+    const blogs = await Blog.find({}).populate('user')
+    response.json(blogs)
   } catch (error) {
     next(error)
   }
 })
 
-blogRouter.post('/', async (request, response, next) => {
-  const { title, url, author, likes, userId } = request.body;
+blogRouter.post('/', userExtractor, async (request, response, next) => {
+  const { title, url, author, likes } = request.body
 
-  const user = await User.findById(userId)
+  if (!request.user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
 
   if (!title || !url) {
-    return response.status(400).json({ error: 'title or url missing' });
+    return response.status(400).json({ error: 'title or url missing' })
   }
-  
-  const newBlog = new Blog({
-    title,
-    url,
-    author, 
-    likes, 
-    user: user._id,
-  })
 
   try {
+    const user = await User.findById(request.user.id)
+    const newBlog = new Blog({
+      title,
+      url,
+      author,
+      likes,
+      user: user._id,
+    })
+
     const savedBlog = await newBlog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -44,8 +44,17 @@ blogRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogRouter.delete('/:id', async (request, response, next) => {
+blogRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== request.user.id.toString()) {
+      return response.status(403).json({ error: 'only the creator can delete this blog' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (error) {
@@ -53,15 +62,19 @@ blogRouter.delete('/:id', async (request, response, next) => {
   }
 })
 
-blogRouter.put('/:id', async (request, response, next) => {
+blogRouter.put('/:id', userExtractor, async (request, response, next) => {
   const { likes } = request.body
 
-  const updatedBlog = await Blog.findByIdAndUpdate(
-    request.params.id,
-    { likes },
-    { new: true }
-  )
-  response.json(updatedBlog)
+  try {
+    const updatedBlog = await Blog.findByIdAndUpdate(
+      request.params.id,
+      { likes },
+      { new: true }
+    )
+    response.json(updatedBlog)
+  } catch (error) {
+    next(error)
+  }
 })
 
 module.exports = blogRouter
